refactor: move fund results polling hooks behind route shell
Fetching and polling are now isolated after pure helpers and presentation were already extracted, reducing risk around the active calculation and comparison refresh behavior.

Constraint: Preserve polling backoff, abort behavior, comparison parsing, query keys, and recalculation refresh fan-out.

Confidence: medium

Scope-risk: moderate

Tested: npx vitest client fund-model-results tests plus fund-results route and comparison contracts

Not-tested: Full release gate
